+++ json_serializable/lib/src/type_helper_ctx.dart
@@ -17,7 +17,10 @@ TypeHelperCtx typeHelperContext(
     TypeHelperCtx._(helperCore, fieldElement, key);
 
 class TypeHelperCtx
-    implements TypeHelperContextWithConfig, TypeHelperContextWithConvert {
+    implements
+        TypeHelperContextWithConfig,
+        TypeHelperContextWithConvert,
+        TypeHelperContextWithEmptyCollectionLogic {
   final HelperCore _helperCore;
   final JsonKey _key;
 
@@ -61,10 +64,23 @@ class TypeHelperCtx
       (TypeHelper th) => th.deserialize(targetType, expression, this));
 
   Object _run(DartType targetType, String expression,
-          Object invoke(TypeHelper instance)) =>
-      _helperCore.allTypeHelpers.map(invoke).firstWhere((r) => r != null,
+      Object invoke(TypeHelper instance)) {
+    _depth++;
+
+    try {
+      return _helperCore.allTypeHelpers.map(invoke).firstWhere((r) => r != null,
           orElse: () => throw UnsupportedTypeError(
               targetType, expression, _notSupportedWithTypeHelpersMsg));
+    } finally {
+      _depth--;
+    }
+  }
+
+  int _depth = 0;
+
+  @override
+  bool get skipEncodingEmptyCollection =>
+      !_key.encodeEmptyCollection && _depth == 1;
 }
 
 final _notSupportedWithTypeHelpersMsg =
@@ -72,6 +88,7 @@ final _notSupportedWithTypeHelpersMsg =
 
 class _ConvertPair {
   static final _expando = Expando<_ConvertPair>();
+
   static _ConvertPair fromJsonKey(JsonKey key) => _expando[key];
 
   final ConvertData fromJson, toJson;