+++ json_serializable/test/kitchen_sink/kitchen_sink.g_any_map__non_nullable__use_wrappers.dart
@@ -31,6 +31,7 @@ class _Factory implements k.KitchenSinkFactory<dynamic, dynamic> {
   bool get nullable => false;
   bool get excludeNull => false;
   bool get explicitToJson => false;
+  bool get noEncodeEmpty => false;
 
   k.KitchenSink ctor({
     int ctorValidatedNo42,