+++ json_serializable/test/shared_config.dart
@@ -16,6 +16,7 @@ final generatorConfigNonDefaultJson =
   createFactory: false,
   createToJson: false,
   disallowUnrecognizedKeys: true,
+  encodeEmptyCollection: false,
   explicitToJson: true,
   fieldRename: FieldRename.kebab,
   generateToJsonFunction: false,