+++ json_serializable/tool/builder.dart
@@ -117,6 +117,8 @@ const _configReplacements = {
   'explicit_to_json':
       _Replacement.addJsonSerializableKey('explicitToJson', true),
   'exclude_null': _Replacement.addJsonSerializableKey('includeIfNull', false),
+  'no_encode_empty':
+      _Replacement.addJsonSerializableKey('encodeEmptyCollection', false),
 };
 
 const _kitchenSinkReplacements = {
@@ -143,6 +145,12 @@ const _kitchenSinkReplacements = {
       'bool get checked => false;',
       'bool get checked => true;',
     )
+  ],
+  'no_encode_empty': [
+    _Replacement(
+      'bool get noEncodeEmpty => false;',
+      'bool get noEncodeEmpty => true;',
+    ),
   ],
   'exclude_null': [
     _Replacement(
@@ -226,6 +234,11 @@ const _fileConfigurationMap = <String, List<List<String>>>{
     ['any_map', 'non_nullable', 'use_wrappers'],
     ['any_map', 'non_nullable'],
     ['any_map'],
+    ['no_encode_empty'],
+    ['no_encode_empty', 'exclude_null', 'use_wrappers'],
+    ['no_encode_empty', 'non_nullable'],
+    ['no_encode_empty', 'exclude_null'],
+    ['no_encode_empty', 'exclude_null', 'non_nullable', 'use_wrappers'],
     ['exclude_null', 'non_nullable'],
     ['exclude_null', 'use_wrappers'],
     ['exclude_null'],